# A few more Enum perf improvements