I started out to just remove the unnecessary boxing allocation that happens when you do enum.ToString("D") or, if the value doesn't map to a named enum entry, just enum.ToString() (there's an internal boxing of the value that happens in addition to any boxing that happens of the enum itself). As part of that, I added a ValueToString method that replaces the GetValue().ToString() calls that appear several times, and in writing that ValueToString, I opted to avoid pinning, instead using Unsafe.As.  Once I did that, I then removed pinning everywhere else it was being done, standardizing on Unsafe.As.

| Method         | Before Time (ns) | After Time (ns) | Time Diff | Before Alloc (b) | After Alloc (b) | Alloc Diff | 
|----------------|------------------|-----------------|-----------|------------------|-----------------|------------| 
|  BoxedHashCode | 6.07             | 5.85            | -3.70%    | 0                | 0               | 0.00%    | 
|    BoxedString | 66.71            | 63.36           | -5.02%    | 56               | 32              | -42.86%    | 
|   BoxedStringD | 30.94            | 26.06           | -15.78%   | 56               | 32              | -42.86%    | 
|   BoxedStringX | 46.91            | 39.21           | -16.41%   | 40               | 40              | 0.00%      | 
|       HashCode | 0.54             | 0.27            | -49.94%   | 0                | 0               | 0.00%    | 
|         String | 74.03            | 67.07           | -9.40%    | 80               | 56              | -30.00%    | 
|        StringD | 34.05            | 28.03           | -17.66%   | 80               | 56              | -30.00%    | 
|        StringX | 52.64            | 41.64           | -20.89%   | 64               | 64              | 0.00%      | 

This is the last of the low-hanging fruit fixes I had for Enum.

cc: @jkotas, @ahsonkhan, @joperezr