@stephentoub 2018-11-30
@dotnet-bot test OSX10.12 x64 Checked Innerloop Build and Test please ("java.nio.channels.ClosedChannelException")